Extract linter_name from payload
[Related Hound pull](https://github.com/houndci/hound/pull/1102)

It is possible for Hound to assign more than one linter to review a single file
in a build. Because of this, when Hound dequeues an incoming
CompletedFileReview, it needs to know both the name of the file linted and the
linter reporting any errors.

This change updates the JShint linter to look for a linter_name key in the
payload object dequeued from Redis. It then passes the linter name back either
upon successful completion of a review or when it detects a configuration error.